Fix scroller logic
The original logic checks if the element is currently scrolling, if so, it will not begin another scroll. However, because it is missing a return, not only will it not begin another scroll, it will also call resetScrolling(). This terminates the ongoing scroll. So even though the element needs scrolling, no scrolling will be done at all. This change adds the missing return.